% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCCqueryString.R
\name{FCCqueryString}
\alias{FCCqueryString}
\title{Compose a query string for an FCC Query Broadcast Station Search}
\usage{
FCCqueryString(
  band = c("FM", "TV", "AM", "fm", "tv", "am"),
  dist = 80,
  dlat2 = 38 + ((54 + (12/60))/60),
  mlat2 = 0,
  slat2 = 0,
  NS = c("N", "S"),
  dlon2 = (-(77 + (36/3600))),
  mlon2 = 0,
  slon2 = 0,
  EW = c("E", "W")
)
}
\arguments{
\item{band}{c('FM', 'TV', 'AM', 'fm', 'tv', 'am')}

\item{dist}{number of kilometers radius in which to search}

\item{dlat2}{degrees latitude}

\item{mlat2}{minutes latitude}

\item{slat2}{seconds latitude}

\item{NS}{c('N', 'S', 'n', 's')}

\item{dlon2}{degrees longitude}

\item{mlon2}{minutes longitude}

\item{slon2}{seconds longitude}

\item{EW}{c('E', 'W', 'e', 'w')}
}
\value{
a character string in the format required by the FCC.
}
\description{
Compose a query string for an FCC Query Broadcast Station Search
}
\examples{
FCCstr1 <- FCCqueryString()
FCCstr2 <- FCCqueryString(dlat2=38L,
  mlat2=54L, slat2=12+1e-9, NS='N',
  dlon2=77, mlon2=0, slon2=36+1e-9, EW='W')
# test integers for dlat and mlat2
# and with real numbers, not integers,
# for slat2 and slon2
}
