% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FCCquery.R
\name{print.FCCquery}
\alias{print.FCCquery}
\title{Print method for FCCquery}
\usage{
\method{print}{FCCquery}(x, file = paste0(deparse(substitute(x)), ".xlsx"), ...)
}
\arguments{
\item{x}{an object of class \code{FCCquery}}

\item{file}{= paste0(deparse(substitute(x)), '.xlsx')}

\item{...}{other arguments for consistency with
the generic \code{print} function.  Not used.}
}
\value{
\code{file}
}
\description{
Print method for objects of class FCCquery.
This creates a file in the working directory
in MS Excel \code{.xlsx} workbook format.
The \code{\link[=data.frame]{data.frame()}} is the first \code{data.frame} tab.
The second tab contains the \code{query} string.
The third tab contains the \code{query_time} in
the format of an object of class \code{\link[=proc.time]{proc.time()}}
giving the time required to extract the
information from \code{fcc.gov}. See @details for
a discussion of the \code{entriesWFmtErrors} tab.
}
\details{
The information in the \code{entriesWFmtErrors} tab
depends on the extent to which the data obtained
from \code{fcc.gov} seemed to be in the standard
format.  If the current algorithm failed to parse
any of the information returned, all that
information (if any) is stored in this
\code{entriesWFmtErrors} tab.  If everything seemed
consistent with expectations, this tab will be
blank.  Otherwise, it will contain a list of
any and all rows that confused the algorithm
with an explanation about what the algorithm did
not understand.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# These examples have taken over a minute to complete
# because of long delays in getting a response from 
# fcc.gov:  
# All FM stations within 80 km of FCC headquarters 
# in Washington, DC
FCCdat1 <- FCCquery()
print(FCCdat1)
\dontshow{\}) # examplesIf}
}
